% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_setup}
\alias{palette_setup}
\title{Setup a color palette}
\usage{
palette_setup(palette = "standard", n, reverse = FALSE)
}
\arguments{
\item{palette}{Character name of palette in palettes_washi. See
\code{names(palettes_washi)} for a list of available palettes.}

\item{n}{Number of colors in palette.}

\item{reverse}{Boolean indicating whether the palette should be reversed.
Default is FALSE.}
}
\value{
A vector of color hex codes.
}
\description{
Choose desired number of colors and whether the colors are reversed.
}
\examples{
palette_setup("color_blind")

palette_setup("green_gradient", 12)
}
