% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_view}
\alias{palette_view}
\title{View a WaSHI palette}
\usage{
palette_view(palette = "color_blind", n, reverse = FALSE)
}
\arguments{
\item{palette}{Character name of palette in palettes_washi. See
\code{names(palettes_washi)} for a list of available palettes.}

\item{n}{Number of colors in palette.}

\item{reverse}{Boolean indicating whether the palette should be reversed.
Default is FALSE.}
}
\value{
A plot with each color displayed.
}
\description{
Show the colors within a palette in a plot.
}
\examples{
palette_view("standard")

palette_view("color_blind")

palette_view("blue_gradient", 4, reverse = TRUE)
}
