% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palettes_washi}
\alias{palettes_washi}
\title{WaSHI color palettes}
\format{
An object of class \code{list} of length 11.
}
\usage{
palettes_washi
}
\description{
Color palettes are stored in a named list.
}
\examples{
# List names of available palettes
names(palettes_washi)

# Get hex codes from a palette using dollar `$name` or double bracket
# `[["name"]]` operators for extracting list elements
palettes_washi$standard

palettes_washi[["green_gradient"]]

# Extract a color from the standard WaSHI palette
palettes_washi[["standard"]][["green"]]
}
\keyword{datasets}
