% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_washi}
\alias{scale_color_washi}
\title{Scales for plotting with washi palettes}
\usage{
scale_color_washi(
  palette = "main",
  reverse = FALSE,
  aesthetics = c("color", "fill"),
  discrete = TRUE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in palettes_washi. See
\code{names(palettes_washi)} for a list of available palettes.}

\item{reverse}{Boolean indicating whether the palette should be reversed.
Default is FALSE.}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. Defaults to
c("color", "fill"), which applies the palette to both the color and fill
aesthetics at the same time.}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or
not. Default is TRUE.}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\value{
A \code{ScaleContinuous} or \code{ScaleDiscrete} object that can be added to a
\code{ggplot} object
}
\description{
Provides compatibility with \code{ggplot2}.
}
\examples{
library(ggplot2)

# Discrete scale
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point(size = 4) +
  theme_minimal() +
  scale_color_washi("main")

# Continuous scale
ggplot(diamonds) +
  geom_point(aes(x = carat, y = price, color = price)) +
  theme_minimal() +
  scale_color_washi("green_gradient", reverse = TRUE, discrete = FALSE)
}
