% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{washi_import_fonts}
\alias{washi_import_fonts}
\title{Import Lato and Poppins fonts}
\usage{
washi_import_fonts()
}
\value{
Imports Lato and Poppins for use in R.
}
\description{
Uses \code{extrafont} and \code{grDevices} to import fonts into the
\code{extrafontdb} database so they may be used in R plots and tables.
This function should only need to be run once. Though \code{library(extrafont)}
must be called in each session.
}
\seealso{
Other font functions: 
\code{\link{washi_install_fonts}()}
}
\concept{font functions}
