% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{washi_pal}
\alias{washi_pal}
\title{WaSHI color palettes}
\format{
An object of class \code{list} of length 11.
}
\usage{
washi_pal
}
\description{
Color palettes are stored in a named list.
}
\examples{
# List names of available palettes
names(washi_pal)

# Get hex codes from a palette using dollar `$name` or double bracket
# `[["name"]]` operators for extracting list elements
washi_pal$standard

washi_pal[["green_gradient"]]

# Extract a color from the standard WaSHI palette
washi_pal[["standard"]][["green"]]
}
\keyword{datasets}
