% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{washi_pal_setup}
\alias{washi_pal_setup}
\title{Setup a color palette}
\usage{
washi_pal_setup(palette = "standard", n, reverse = FALSE)
}
\arguments{
\item{palette}{Character name of palette in washi_pal. See
\code{names(washi_pal)} for a list of available palettes.}

\item{n}{Number of colors in palette.}

\item{reverse}{Boolean indicating whether the palette should be
reversed. Default is FALSE.}
}
\value{
A vector of color hex codes.
}
\description{
Choose desired number of colors and whether the colors are
reversed.
}
\examples{
washi_pal_setup("color_blind")

washi_pal_setup("green_gradient", 12)
}
\seealso{
Other color palette functions: 
\code{\link{washi_pal_view}()},
\code{\link{washi_pal}}
}
\concept{color palette functions}
