% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{washi_register_fonts}
\alias{washi_register_fonts}
\title{Register Lato and Poppins fonts in R}
\usage{
washi_register_fonts()
}
\value{
Registers Lato and Poppins for use in R.
}
\description{
Uses \code{extrafont} and \code{grDevices} to register fonts so they may be
used in R plots and tables. This function only needs to be run
once. Though \code{library(extrafont)} must be called in each session.
}
\seealso{
Other font functions: 
\code{\link{washi_install_fonts}()}
}
\concept{font functions}
