% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{washi_theme}
\alias{washi_theme}
\title{Create standard washi plots}
\usage{
washi_theme(
  font_title = "Lato Black",
  color_title = "#151414",
  font_subtitle = "Lato Black",
  color_subtitle = "#151414",
  font_caption = "Poppins",
  color_caption = "#151414",
  font_plot = "Poppins",
  color_plot_text = "#151414",
  text_scale = 1,
  legend_position = "top",
  facet_space = 2,
  color_gridline = washi_pal[["standard"]][["tan"]],
  gridline_y = TRUE,
  gridline_x = TRUE,
  ...
)
}
\arguments{
\item{font_title}{Font family for title. Defaults to "Lato Black".}

\item{color_title}{Font color for title. Defaults to almost black.}

\item{font_subtitle}{Font family for subtitle. Defaults to "Lato
Black".}

\item{color_subtitle}{Font color for title. Defaults to almost
black.}

\item{font_caption}{Font family for caption. Defaults to "Poppins".}

\item{color_caption}{Font color for title. Defaults to almost
black.}

\item{font_plot}{Font family for plot text. Defaults to "Poppins".}

\item{color_plot_text}{Font color for title. Defaults to almost
black.}

\item{text_scale}{Scalar that will grow/shrink all text defined
within.}

\item{legend_position}{Position of legend ("none", "left", "right",
"bottom", "top", or two-element numeric vector). Defaults to
"top".}

\item{facet_space}{Controls how far apart facets are from each
other.}

\item{color_gridline}{Gridline color. Defaults to washi tan.}

\item{gridline_y}{Boolean indicating whether major gridlines are
displayed for the y axis. Default is TRUE.}

\item{gridline_x}{Boolean indicating whether major gridlines are
displayed for the x axis. Default is TRUE.}

\item{...}{Pass any parameters from theme that are not already
defined within}
}
\description{
Adapted from
\href{https://github.com/USAID-OHA-SI/glitr/blob/main/R/si_style.R}{glitr::si_style()}.
All changed defaults from this function can be overriden by another
call to \code{\link[ggplot2:theme]{ggplot2::theme()}} with the desired changes.
}
\examples{
\dontrun{
library(extrafont)
library(ggplot2)

# Single plot

ggplot(iris, aes(Sepal.Length, y = Sepal.Width, color = Species)) +
  labs(
    title = "washi_theme() + washi_scale()",
    subtitle = "This is a subtitle.",
    caption = "This is a caption."
  ) +
  geom_point(size = 2.5) +
  washi_theme() +
  washi_scale()

# Facet plot

ggplot(iris, aes(x = Sepal.Length, fill = Species)) +
  labs(
    title = "washi_theme() + washi_scale()",
    subtitle = "Example of facet_grid().",
    caption = "This is a caption."
  ) +
  geom_histogram(bins = 15) +
  facet_grid(. ~ Species) +
  washi_theme(legend_position = "none") +
  washi_scale()
}
}
