% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{washi_install_fonts}
\alias{washi_install_fonts}
\title{Install Lato and Poppins fonts on your system}
\usage{
washi_install_fonts()
}
\value{
Opens a directory containing font files.
}
\description{
Opens the font files to install on your computer.
}
\concept{font functions}
