% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{washi_register_fonts}
\alias{washi_register_fonts}
\title{Register Lato and Poppins fonts in R}
\usage{
washi_register_fonts()
}
\description{
Register Lato and Poppins fonts in R
}
\seealso{
Other font: 
\code{\link{washi_install_fonts}()}
}
\concept{font}
